+++ src/asteroid.rs
@@ -12,7 +12,7 @@ use physics::{delete_entity, Position, Velocity};
 use rand::prelude::*;
 use specs::{Component, Entities, Read, Join, LazyUpdate, NullStorage,
             ReadStorage, System};
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 
 /// An asteroid
 #[derive(Default)]
@@ -74,8 +74,8 @@ impl<'a> System<'a> for SysAsteroid {
             let bi = b as i32 + 1;
             for y in -ai..ai {
                 for x in -bi..bi {
-                    let x = x as f64;
-                    let y = y as f64;
+                    let x = x as f32;
+                    let y = y as f32;
                     if x * x * a * a + y * y * b * b <= a * a * b * b {
                         blocks.push(([x, y], Block::new(BlockInner::Rock)));
                     }
